Bug: Added Controller to abstract functionalities
