+++ app/routes/index.js
@@ -2,7 +2,7 @@ import route from './entry_routes';
 
 const entryRoutes = route;
 
-export default function router(app, array) {
-  entryRoutes(app, array);
+export default function router(app) {
+  entryRoutes(app);
   // Other route groups could go here, in the future
 }
