+++ server.js
@@ -1,12 +1,12 @@
 import express from 'express';
 import bodyParser from 'body-parser';
-import { Pool } from 'pg';
-import dotenv from 'dotenv';
+// import { Pool } from 'pg';
+// import dotenv from 'dotenv';
 import router from './app/routes';
 
-dotenv.config();
-const dbConfig = { connectionString: process.env.DATABASE_URL };
-const pool = new Pool(dbConfig);
+// dotenv.config();
+// const dbConfig = { connectionString: process.env.DATABASE_URL };
+// const pool = new Pool(dbConfig);
 
 // create express app
 const app = express();
@@ -14,7 +14,8 @@ const app = express();
 
 app.use(bodyParser.json());
 app.use(bodyParser.urlencoded({ extended: true }));
-router(app, pool);
+// app.use('/v1', routes)(app, pool);
+router(app);
 // require('./app/routes')(app, array);
 
 // listen for requests
@@ -23,4 +24,4 @@ const server = app.listen(process.env.PORT || 5000, () => {
 });
 
 // module.exports = [server, array];
-export { server, pool };
+export default { server };
