+++ test/login.spec.js
@@ -1,6 +1,6 @@
 import chai from 'chai';
 import chaiHttp from 'chai-http';
-import { server, pool } from '../server';
+import { server } from '../server';
 
 const { expect } = chai;
 chai.use(chaiHttp);
